chore: moved modal to outside, class to classname and some tailwind fix
