+++ client/src/admin/components/Modal.js
@@ -1,14 +1,14 @@
-import { useGlobalAdminAppContext } from '../../context/adminAppContext';
+import { useGlobalAdminAppContext } from '../context/adminAppContext';
 import { useRef, useEffect, useState } from 'react';
 import { RiCloseLine } from 'react-icons/ri';
 import moment from 'moment';
 
 import 'react-datepicker/dist/react-datepicker.css';
-import styles from '../../Modal.module.css';
+import styles from '../Modal.module.css';
 
-import { Loading } from '../Screens';
-import { ConfirmModalForData } from './Confirmation';
-import { SubmitButton, BusyButton } from '../Buttons';
+import { Loading } from './Screens';
+import { ConfirmModalForData } from './superadmin/Confirmation';
+import { SubmitButton, BusyButton } from './Buttons';
 
 const AddModal = ({
   setShowAddModal,
@@ -479,7 +479,7 @@ const AddModal = ({
                           Search
                         </label>
                         <div className='relative'>
-                          <div class='flex absolute inset-y-0 left-0 items-center pl-3 pointer-events-none'>
+                          <div className='flex absolute inset-y-0 left-0 items-center pl-3 pointer-events-none'>
                             <svg
                               aria-hidden='true'
                               className='w-5 h-3 text-gray-500 dark:text-gray-400'
@@ -510,7 +510,7 @@ const AddModal = ({
                           {searching === false ? (
                             <button
                               type='button'
-                              class='text-white absolute right-2.5 bottom-2.5 bg-admin3 hover:bg-admin4 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2'
+                              className='text-white absolute right-2.5 bottom-2.5 bg-admin3 hover:bg-admin4 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2'
                               onClick={async (e) => {
                                 e.preventDefault();
                                 setSearching(true);
